fix: Can not enable proxy for non-root user (#21)
The plugin tries to enable apt proxy, but it is not support to use apt for non-root users, so as the proxy for apt. I only determine if the current user has sudo privileges. If the user does not have sudo privileges, the proxy will not be set for apt.